requirements.txt
flask
opencv-python-headless
numpy
Pillow
qrcode
